+++ src/Documentation/BaseEntity.ts
@@ -1,3 +1,5 @@
+import { Constructor } from "@/Documentation/MetadataTypes";
+
 /**
  * The base class for entities defined in your application.
  *
@@ -8,11 +10,9 @@
 export default class BaseEntity {
   /**
    * Constructor used to create a new Entity and fill it with data.
-   *
-   * @param [options] A map of properties with values that can be filled in the current entity.
    */
-  public constructor(options?: Record<string, unknown>) {
-    this.init(options);
+  public constructor() {
+    //NOP
   }
 
   /**
@@ -20,12 +20,27 @@ export default class BaseEntity {
    *
    * @param [options] A map of properties with values that can be filled in the current entity.
    */
-  public init(options?: Record<string, unknown>) {
+  public init(options?: Record<string | number | symbol, any>) {
     if (options) {
-      Object.keys(options).forEach((key) => {
-        this[key] = options[key];
-      });
+      Object.assign(this, options);
     }
+
+    return this;
+  }
+
+  /**
+   * Creates a clone of the current entity.
+   *
+   * @param [properties] A list of properties that you want to pass to the new instance.
+   */
+  public clone<T = BaseEntity>(properties?: Record<string, any>): T {
+    const copy = new (this.constructor as { new (): T })();
+    Object.assign(copy, this);
+
+    if (properties && Object.keys(properties).length > 0) {
+      Object.assign(copy, properties);
+    }
+    return copy;
   }
 
   /**
@@ -33,8 +48,8 @@ export default class BaseEntity {
    *
    * Use this method when you want to extract the data from the entity.
    */
-  public toJSON<T = Record<string, unknown>>(): T {
-    return JSON.parse(JSON.stringify(this));
+  public toJSON<T = Record<string | number | symbol, any>>(): T {
+    return Object.assign({}, this) as T;
   }
 
   /**
@@ -43,7 +58,20 @@ export default class BaseEntity {
    * Use this method when you want to extract the data from the entity.
    * This is an alias for the `toJSON` method.
    */
-  public toObject<T = Record<string, unknown>>(): T {
+  public toObject<T = Record<string | number | symbol, any>>(): T {
     return this.toJSON<T>();
   }
+}
+
+/**
+ * Function used to create a new entity with data.
+ *
+ * @param EntityClass The Entity we want to initialize.
+ * @param options A map of properties with values that can be filled in the current entity.
+ */
+export function createEntity<Entity extends BaseEntity>(EntityClass: Constructor<Entity>, options?: Record<string | symbol | number, any>): Entity {
+  const instance = new EntityClass();
+  instance.init(options);
+
+  return instance;
 }